added release and debug compile version, better clean up

+++ Makefile
@@ -1,15 +1,31 @@
-C_FLAGS =-Wall -Wextra  -fsanitize=address -g
+CC = gcc
+C_FLAGS = -Wall -Wextra
+C_EXFLAGS = 
 
 c_vector_src_path := $(or $(C_VECTOR_ROOT), .)
 c_vector.c := $(c_vector_src_path)/c_vector.c
 
-all: c_vector.o
+all: release
+
+DEBUG_FLAGS = -O0 -fsanitize=address -g
+RELEASE_FLAGS = -O2
+
+debug: C_FLAGS += $(DEBUG_FLAGS)
+debug: clean compile
+
+release: C_FLAGS += $(RELEASE_FLAGS)
+release: clean compile
+
+
+compile: c_vector.o
 
 c_vector.o: $(c_vector.c)
-	gcc $(C_FLAGS) $(c_vector.c) -c
+	$(CC) $(C_FLAGS) $(C_EXFLAGS) $(c_vector.c) -c
 
 c_vector_clean:
+ifeq ($(wildcard c_vector.o), c_vector.o)
 	rm c_vector.o
+endif
 
 
 clean: c_vector_clean
